EOS-20329: CLI Framework movement (#463)
* CLI Framework movement. initial commit


* Cortxcli framework movement to py-utils


* Remove core framework part from csm


* Porting validators as cli validators


* Adding type_target for non builtin datatype


* Adding type_target for non builtin datatype


* Modifying unittestcases


* Code cleanup and codacy issue resolved


* Modifications for cli_setup


* Temp checkin


* Revert temprary change

* Add liscence

* Add common validators to py-utils


* Chaanged dir name cli to cli_framework


* Changing ArgumentError->CliError


* Addressing comment


* Minor modification


* Addressing comments


* Addressing comments


* EOS-20752: Renaming BaseError class name to UtilsError (#476)



* EOS-20537: Unhandled exception after crossing S3 accounts limit (#475)

* EOS-20347 PAM AUTH POC (#464)

* Correction in csm_agent login url and other minor changes


* Curl option for failed API response has set to TRUE


* PAM readme.updated


* Password prompt issue resolved


* PAM module renamed to pam_cortx


* PAM readme updated


* Removed unnecessary comments


* Trying to fix codacy issue



* EOS-20537: add Max{Account,User}LimitExceeded to the enum of known IamErrors



* Delete .whitesource (#472)


* EOS-20352: BE:Admin to delete S3 account (#480)

* EOS-20352: grant admin permission to delete S3 accounts


* EOS-20352: allow sgiamadmin to delete S3 accounts


* EOS-20352: fix admin's session termination



* EOS-20700: API to get all the unsupported features (#477)

* EOS-20700: API to get all the unsupported features


* Addressed codacy comments

* Addressed codacy

* Addressed codacy issues

* Removed code for removing lyve_pilot from permissions


* Codacy issues addressed


* EOS-20343: Cleaning R1 code for health implementation (#485)

* EOS-20347 PAM AUTH POC (#464)

* Correction in csm_agent login url and other minor changes


* Curl option for failed API response has set to TRUE


* PAM readme.updated


* Password prompt issue resolved


* PAM module renamed to pam_cortx


* PAM readme updated


* Removed unnecessary comments


* Trying to fix codacy issue



* t-main updation (#484)


* EOS-20343: Cleaning R1 code for health implementation


* EOS-20343: Cleaning R1 code for health implementation


* Fixing codacy issues


* EOS-20345: USL-Setup integration (#490)

* Minor modifications for usl-setup


* Update setup.yaml


* Return 0 on success


* EOS-20670: Create new endpoint that returns audit logs schema (#473)

* EOS-20670: Implement stub method handler for audit logs schema info


* EOS-20670: Implement method view for audit logs schema info


* Revert "EOS-20752: Renaming BaseError class name to UtilsError (#476)" (#493)

This reverts commit fcc418b531c20a0b79a9e866297ccbd9e7001a07.

* EOS-15573: Fixed an issue if an alert comes without corresponding fault, 'fault resolve' Alert appears as unresolved (#491)

* EOS-20347 PAM AUTH POC (#464)

* Correction in csm_agent login url and other minor changes


* Curl option for failed API response has set to TRUE


* PAM readme.updated


* Password prompt issue resolved


* PAM module renamed to pam_cortx


* PAM readme updated


* Removed unnecessary comments


* Trying to fix codacy issue



* t-main updation (#484)


* EOS-15573: Fixed an issue if an alert comes without corresponding fault, 'fault resolve' Alert appears as unresolved



* EOS-20356: Reset command for csm setup (#487)

* Reset command for csm setup


* Minor modifications


* Delete unsupported feature entry for csm


* Addressing comments


* minor modifications



* EOS-20355: Cleanup command for csm_setup (#479)

* EOS-20347 PAM AUTH POC (#464)

* Correction in csm_agent login url and other minor changes


* Curl option for failed API response has set to TRUE


* PAM readme.updated


* Password prompt issue resolved


* PAM module renamed to pam_cortx


* PAM readme updated


* Removed unnecessary comments


* Trying to fix codacy issue



* Cleanup command for csm_setup


* Modifications for cleanup command


* t-main updation (#484)


* Modifications for Cleanup command for csm-setup


* Modifications for Cleanup command for csm-setup


* Cleanup cmd modifications


* Minor modification


* Addressing comments


* Resolve conflicts



* EOS-20530: CLI for Audit logs (#488)

* EOS-20530: added CLI for CSM logs


* CLI added for s3 audit logs


* date changes and validations


* added date changes for s3 audit logs


* added unit test cases for csm audit logs cli



* EOS-20327: Audit logs sorting on text fields (#462)

* EOS-20488: Make ES mappings for `csmauditlogs` match CSM business logic


* EOS-20487: Remove restrictions on sortable fields


* EOS-20487: Clean up imports


* EOS-20487: Remove stale field `msg` from audit log model


* EOS-20487: Fix response code type on audit log model



* Adding auditlog changes.


* EOS-20901: Audit logs are not displayed for today in CLI (#494)

* EOS-20530: added CLI for CSM logs


* CLI added for s3 audit logs


* date changes and validations


* added date changes for s3 audit logs


* added unit test cases for csm audit logs cli


* Fix for audit logs for today


* Addressing comments


* Addressing comments


* EOS-20354: Preupgrade and Postupgrade commands for csm-setup and USL-Setup (#492)

* Backup command for csm_setup


* Backup command for csm_setup


* Post_upgrade cmd for csm_setup


* Post_upgrade cmd for csm_setup


* Change  Backup to Preinstall


* Change  Backup to Preupgrade


* Modifications for Post-upgrade cmd for csm-setup


* Modifications for Post-upgrade cmd for csm-setup


* Preugrade and Postupgrade commands for USL setup


* Minot modifications


* minor modification

* Adding templates for preupgrade and postupgrade


* Adding pre and post upgrade entires to setup.yam


* Adding pre and post upgrade entires to setup.yam


* Addressing comments


* Addressing comments


* Added stub classes and methods for Health API. (#497)



* EOS-20999:Name and Description updated for change password command (#499)



* Adding type_method


* Addressing comments


* Update csm_agent.py

* Update csm_client.py
