+++ csm/cli/schema/s3accounts.json
@@ -65,7 +65,9 @@
           {
             "flag": "-p",
             "dest": "password",
-            "type": "Terminal.get_password",
+            "type": "str",
+            "type_method": "Terminal.get_password",
+            "type_target": "cortx.utils.cli_framework.terminal",
             "help": "",
             "default": "",
             "suppress_help": true,
@@ -133,7 +135,9 @@
           {
             "flag": "-p",
             "dest": "password",
-            "type": "Terminal.get_password",
+            "type": "str",
+            "type_method": "Terminal.get_password",
+            "type_target": "cortx.utils.cli_framework.terminal",
             "help": "",
             "default": "",
             "suppress_help": true,
