+++ csm/cli/schema/s3buckets.json
@@ -50,7 +50,9 @@
       "args": [
         {
           "flag": "bucket_name",
-          "type": "Validatiors.bucket_name",
+          "type": "str",
+          "type_method": "Validators.file_parser",
+          "type_target": "csm.common.payload",
           "help": "Give a bucket name to create.",
           "json": true
         },
