+++ csm/cli/schema/s3iamusers.json
@@ -60,7 +60,9 @@
         {
           "flag": "-passwd",
           "dest": "password",
-          "type": "Terminal.get_password",
+          "type": "str",
+          "type_method": "Terminal.get_password",
+          "type_target": "cortx.utils.cli_framework.terminal",
           "help": "",
           "default": "",
           "suppress_help": true,
@@ -125,7 +127,9 @@
         {
           "flag": "-p",
           "dest": "password",
-          "type": "Terminal.get_password",
+          "type": "str",
+          "type_method": "Terminal.get_password",
+          "type_target": "cortx.utils.cli_framework.terminal",
           "help": "",
           "default": "",
           "suppress_help": true,
