+++ csm/common/converter.py
@@ -1,5 +1,5 @@
 # CORTX-CSM: CORTX Management web and CLI interface.
-# Copyright (c) 2020 Seagate Technology LLC and/or its Affiliates
+# Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published
 # by the Free Software Foundation, either version 3 of the License, or
@@ -12,33 +12,17 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
-from csm.core.blogic import const
-from cortx.utils.cli_framework.command_factory import CommandFactory
-from argparse import ArgumentError
-import unittest
-import json
-import os
-from csm.test.common import Const
 
-permissions = {
-            'security': { 'read': True }
-}
+import datetime
 
-show_command = CommandFactory.get_command(
-    ["security","details"], permissions)
-t = unittest.TestCase()
 
-def test_1(*args):
-    expected_output = 'security'
-    actual_output = show_command.name
-    t.assertEqual(actual_output, expected_output)
+class Converter:
 
-def test_2(*args):
-    expected_output = 'get'
-    actual_output = show_command.method
-    t.assertEqual(actual_output, expected_output)
+    @staticmethod
+    def convert_date_to_epoch(date):
+        return int(datetime.datetime.strptime(str(date), '%Y-%m-%d').timestamp())
 
-def init(args):
-    pass
-
-test_list = [test_1, test_2]
+    @staticmethod
+    def convert_end_date_to_epoch(date):
+        return int(datetime.datetime.strptime(str(date), '%Y-%m-%d').replace(
+                    hour=23, minute=59, second=59).timestamp())
