+++ csm/conf/csm_setup.py
@@ -38,6 +38,7 @@ class CsmSetupCommand:
 
     def _validate(self):
         ''' Validate setup command '''
+
         if len(self._args) < 2:
             raise Exception('Usage: csm_setup -h')
 
@@ -48,7 +49,7 @@ class CsmSetupCommand:
         # hardcoded permissions
         csm_setup_permissions_dict = {'update': True}
         cmd_obj = CommandParser(Json(const.CSM_SETUP_FILE).load(), csm_setup_permissions_dict)
-        cmd_obj.handle_main_parse(subparsers)
+        cmd_obj._handle_main_parse(subparsers)
         namespace = parser.parse_args(self._args)
         sys_module = sys.modules[__name__]
         for attr in ['command', 'action', 'args']:
@@ -59,7 +60,7 @@ class CsmSetupCommand:
     def process(self):
         ''' Parse args for csm_setup and execute cmd to print output '''
         self._cmd = self._get_command()
-        obj = CsmDirectClient()
+        obj = CliClient()
         response = self._loop.run_until_complete(obj.call(self._cmd))
         if response:
             self._cmd.process_response(out=sys.stdout, err=sys.stderr,
@@ -70,10 +71,10 @@ if __name__ == '__main__':
     sys.path.append(os.path.join(os.path.dirname(pathlib.Path(os.path.realpath(__file__))), '..', '..'))
     from cortx.utils.conf_store.conf_store import Conf
     from csm.common.payload import *
-    from csm.cli.command import CommandParser
+    from cortx.utils.cli_framework.parser import CommandParser
     from csm.core.blogic import const
     from cortx.utils.log import Log
-    from csm.cli.csm_client import CsmDirectClient
+    from cortx.utils.cli_framework.client import CliClient
     from csm.conf.post_install import PostInstall
     from csm.conf.prepare import Prepare
     from csm.conf.configure import Configure
